Vizualicação de locais cadastrados

+++ Back-end/api-sangue-corinthiano/api-sangue-corinthiano/src/main/kotlin/api-sistema/school.sptech.api/repository/InstituicaoRepository.kt
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
 
 @Repository
-interface InstituicaoRepository : JpaRepository<Instituicao, Long> {
+interface InstituicaoRepository : JpaRepository<Instituicao, Int> {
 
     @Query("""
         SELECT new ProjetoPI.ProjetoDoadores.dto.InstituicaoDto(i.nome, e.latitude, e.longitude)
